Add MySafeArea to ensure the bottom padding is standarize between screen when we run the application under Web (JS/Wasm).
Clean some code as it being auto formatted by dart when the vscode settings is to auto format document when save/editiing is turn on to make it easier to see and check based on my preferences.

Change budget transaction page from using old _isLoading method to FutureBuilder to ensure we are align with the way application getting the data from backend as FutureBuilder is easier to control whether it's success, got error, or loading.