Add summary per category on the transaction search. So user can check each summary expense/income/transfer based on the transaction category, which sometimes more usable than name, example we want to know how many times we are eating out, instead of how many times we go to McDonalds.
Revamp the look on the transaction search summary also for the expanded child, instead add the category, as the child is represent the summary of that name/category in that year, we just need to show the year and the amount instead fully draw the item with icon, etc.

Change the suffix behaviour on the search text fields to instead submit the transaction (as we already have on submit), and icon button to submit the search, change the suffix to clear the text instead.

Add delete icon button on the chip category selection, this is to make it obvious that to delete the category you will just need to tap on the chip, instead press clear and redo the category selection.