Add Monthly and Yearly selection on wallet stat page.
Change BarStat to automatically showed the difference amount between income and expense on that month/year, instead just showed as indicator only.
This is already showed in wallet stat, so we put it also on BarStat widget as we will use the same widget in wallet stat page to align the design between each statistic pages.

Change income and expense variable to not null, and handle the null check in the page that using BarStat widget instead.